refactor(crazygames): move nav-button updater out of Main.ts
Extract updateCrazyGamesNavButton (~90 lines) into
CrazyGamesAccountButton.ts verbatim; Main.ts just imports it.
